Moved IfFormik to components because it is reuseable.
