+++ meteor-app/imports/ui/pages/register/register-inner-form.jsx
@@ -4,8 +4,8 @@ import { Form } from "formik";
 import i18n from "meteor/universe:i18n";
 
 import withUpdateOnChangeLocale from "/imports/ui/hoc/update-on-change-locale.jsx";
+import IfFormik from "/imports/ui/components/if-formik.jsx";
 
-import IfFormik from "./if-formik.jsx";
 import RoleInput from "./role-input.jsx";
 import FormGroup from "./form-group.jsx";
